+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/index.ts
@@ -1,6 +1,6 @@
 import type { OpenAPI } from '../../../openapi/index.js'
 import { componentsCode } from './components/index.js'
-import { routeCode } from './route/route-code.js'
+import { routeCode } from './route/index.js'
 
 /**
  * Generates Hono-compatible TypeScript code from an OpenAPI object.
